WIP Reference the values files from the local git repo
When deploying Industrial Edge with the operator, which sets
ValuesDirectoryURLs [1] to an empty string, the manuela-tst application
will fail with the following error:

    global.namespace=$ARGOCD_APP_NAMESPACE --set
    global.pattern=industrial-edge --set
    global.hubClusterDomain=apps.bandini-dc.blueprints.rhecoeng.com --set
    global.localClusterDomain=apps.bandini-dc.blueprints.rhecoeng.com --set
    global.valuesDirectoryURL= --post-renderer ./kustomize` failed exit
    status 1: Error: open /values-global.yaml: no such file or directory

The reason for this is that the operator sets ValuesDirectoryURLs to '' [1]
and with this plugin the helm template command will fail to find those
values files. When the plugin runs, its folder is the chart folder itself
(charts/datacenter/manuela-tst in this case) and the values files are in
the root folder. So let's call "git rev-parse --show-toplevel" to get
the full path of the values files.

Tested on an IE environment deployed via operator and now manuela-tst
works correctly.

Note this PR depends on [3].

Co-Authored-By: Lester Claudio <claudiol@redhat.com>

[1] https://github.com/hybrid-cloud-patterns/patterns-operator/commit/ba3c35a15dd5f239c7a7d337747b01d494278429
[2] https://github.com/hybrid-cloud-patterns/common/pull/98
[3] This depends on PR#111