Switch to using ./subcharts as the unpacked folder for subcharts that need patching
Previously we always fetch the upstream helm chart, patched the tgz and
then placed it under ./charts. At this point in time the subchart
dependency had "repository: https://charts.external-secrets.io" in the
dependencies list in the top-level Chart.yaml file. This was all good
(albeit a bit cumbersome) until we tried switching our charts to live
in an external repository.

When the chart (e.g. golang-external-repository) lives in a remote
helm repository, helm always decides to refetch the remote chart and
ignores the fact that the chart already exists (and in a patched form)
locally under ./charts.

To avoid this issue, we point the dependency to a local "repository"
folder called "subcharts/" that contains the *extracted* sources. This
makes it so that "helm dependency *" commands still work (they would not
if we place the extracted sources directly under charts/, hence the
separate name).

As a bonus we now updating a chart is a bit simpler and the README.md
reflects that.